+++ tests/parser/auto-resolve/auto-resolve.test.ts
@@ -8,7 +8,6 @@ test.each([`a b c?`, `a b c*`, `a b c+`, `a? b c d?`])(
         .define({
           test: gr,
         })
-        .expand()
         .entry("test")
         .build(new Lexer.Builder().build(), { checkConflicts: true });
     }).not.toThrow(`Unresolved R-S conflict`);